BRCD-4513 enable_client_plugin script copies view only if there is a plugin

+++ docker/billrun-docker/billrun-php74/enable_client_plugin.sh
@@ -1,17 +1,19 @@
 #!/bin/bash
 
-if test -d "/plugin/application/plugins/"; then
+if test -d "/plugin/application/"; then
+     if test -d "/plugin/application/plugins/"; then
      cd /plugin/application/plugins/
-     for f in *.php
-     do
-          rm -f "/billrun/application/plugins/"$f
-          ln -s /plugin/application/plugins/$f "/billrun/application/plugins/"$f
-     done
-     for f in /plugin/application/plugins/*.json
-     do
-          [ -f "$f" ] || break
-          echo "configuration.include[] = $f" >> /billrun/conf/container.ini
-     done
+         for f in *.php
+         do
+              rm -f "/billrun/application/plugins/"$f
+              ln -s /plugin/application/plugins/$f "/billrun/application/plugins/"$f
+         done
+         for f in /plugin/application/plugins/*.json
+         do
+              [ -f "$f" ] || break
+              echo "configuration.include[] = $f" >> /billrun/conf/container.ini
+         done
+     fi
      for f in /plugin/conf/*.json
      do
           [ -f "$f" ] || break
@@ -30,7 +32,6 @@ fi
 
 if test -d "/plugin/conf/translations/overrides/"; then
     rm -rf /billrun/conf/translations/overrides/
-    mkdir -p /billrun/conf/translations/
     ln -s /plugin/conf/translations/overrides /billrun/conf/translations/overrides
 fi
